+++ build-js.sh
@@ -144,6 +144,7 @@ echo "building shared bitcode"
   $BINARYEN_SRC/passes/SSAify.cpp \
   $BINARYEN_SRC/passes/StackIR.cpp \
   $BINARYEN_SRC/passes/Strip.cpp \
+  $BINARYEN_SRC/passes/StripTargetFeatures.cpp \
   $BINARYEN_SRC/passes/TrapMode.cpp \
   $BINARYEN_SRC/passes/Untee.cpp \
   $BINARYEN_SRC/passes/Vacuum.cpp \