+++ src/passes/Print.cpp
@@ -1456,7 +1456,6 @@ Pass* createPrintStackIRPass() {
 
 std::ostream& WasmPrinter::printModule(Module* module, std::ostream& o) {
   PassRunner passRunner(module);
-  passRunner.setFeatures(FeatureSet::All);
   passRunner.setIsNested(true);
   passRunner.add<Printer>(&o);
   passRunner.run();