+++ src/tools/wasm-as.cpp
@@ -91,9 +91,11 @@ int main(int argc, const char *argv[]) {
     Fatal() << "error in parsing input";
   }
 
+  wasm.features = FeatureSet::All;
+
   if (options.extra["validate"] != "none") {
     if (options.debug) std::cerr << "Validating..." << std::endl;
-    if (!wasm::WasmValidator().validate(wasm, FeatureSet::All,
+    if (!wasm::WasmValidator().validate(wasm,
          WasmValidator::Globally | (options.extra["validate"] == "web" ? WasmValidator::Web : 0))) {
       WasmPrinter::printModule(&wasm);
       Fatal() << "Error: input module is not valid.\n";