+++ src/wasm.h
@@ -924,6 +924,9 @@ public:
   std::vector<UserSection> userSections;
   std::vector<std::string> debugInfoFileNames;
 
+  FeatureSet features = FeatureSet::MVP;
+  bool emitFeaturesSection = false;
+
   MixedArena allocator;
 
 private: